feat: attach to existing cluster/child processes
Simple enough. We use the process enumeration that we previously
grabbed from the old node-debug, and spawn a watchdog for each child
that we see. The watchdog will then hook these into the existing IPC
server we set up, and they come in as new child targets.

Fixes #132

+++ src/targets/node/watchdog.ts
@@ -17,7 +17,7 @@ process.on('unhandledRejection', e => debugLog(`Unhandled rejection: ${e}`));
 
 process.on('exit', () => {
   debugLog('KILL');
-  if (info.pid) {
+  if (info.pid && !info.dynamicAttach) {
     process.kill(Number(info.pid));
   }
 });
